refactor: split interfaces into appropriate files
