+++ src/App/types/column.ts
@@ -1,0 +1,8 @@
+export interface IColumnModalProps {
+	isOpen: boolean;
+	handleClose: () => void;
+	currentTitle: string;
+	currentId: string;
+	currentBoardId: string;
+	currentOrder: number;
+}
